refactor: 관심사 분리를 위해 새로운 클래스 작성
- 컴퓨터의 랜덤한 번호 생성, 힌트 생성 등의 로직 처리를 위한 NumberBaseballManager 클래스 작성
- 기존 App 클래스에서 필요없는 코드 제거
